feat: layout gnb 없는 버전 추가
