+++ src/pages/Splash/splash.module.scss
@@ -1,0 +1,4 @@
+.splash {
+  display: flex;
+  flex-direction: column;
+}
