+++ src/routes/index.tsx
@@ -4,11 +4,12 @@ import Home from 'pages/Home'
 import Profile from 'pages/Profile'
 import Signup from 'pages/Signup'
 import Login from 'pages/Login'
-import Layout from 'components/Layout'
+import { Layout, LayoutGnb } from 'components/Layout'
 import NotFound from 'pages/NotFound'
 import { PostDetail, PostEdit, PostWrite } from 'pages/Post'
 import MyPost from 'pages/MyPost'
 import Bookmark from 'pages/Bookmark'
+import Splash from 'pages/Splash'
 import styles from './routes.module.scss'
 
 const RootRoute = () => {
@@ -16,9 +17,12 @@ const RootRoute = () => {
     <div className={styles.app}>
       <Routes>
         <Route path='/' element={<Layout />}>
+          <Route path='/login' element={<Login />} />
+        </Route>
+        <Route path='/' element={<LayoutGnb />}>
+          <Route path='/splash' element={<Splash />} />
           <Route path='/' element={<Home />} />
           <Route path='/signup' element={<Signup />} />
-          <Route path='/login' element={<Login />} />
           <Route path='/profile' element={<Profile />} />
           <Route path='/mypost' element={<MyPost />} />
           <Route path='/bookmark' element={<Bookmark />} />
